add use-pagination hook
